% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apisniffer-package.R
\docType{package}
\name{apisniffer-package}
\alias{apisniffer}
\alias{apisniffer-package}
\title{apisniffer: Find APIs Called by Websites}
\description{
Use web traffic to locate APIs invoked by websites.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/apisniffer}
  \item \url{https://jonthegeek.github.io/apisniffer/}
  \item Report bugs at \url{https://github.com/jonthegeek/apisniffer/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
