% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff.R
\name{sniff}
\alias{sniff}
\title{Detect API calls in a web page}
\usage{
sniff(url, ..., min_wait = 8, max_wait = 60)
}
\arguments{
\item{url}{The URL to load.}

\item{...}{These dots are for future extensions and must be empty.}

\item{min_wait}{The minimum number of seconds to wait for API calls to begin.
The default 8 seconds worked for the 1 page I have tested so far.}

\item{max_wait}{The maximum number of seconds to wait for API calls to
complete. If there are still calls happening when this time is reached,
they will be stopped and ignored.}
}
\value{
A messy list that will eventually tell you useful information about
API calls.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Load a page, wait for API calls to
complete, and return data about the results.
}
